creation

+++ newpolka/pk_cherni.c
@@ -457,7 +457,7 @@ void cherni_backsubstitute(pk_internal_t* pk, matrix_t* con, size_t rank)
   for (k=(int)rank-1; k>=0; k--) {
     j = pk->cherni_intp[k];
     for (i=0; i<con->nbrows; i++) {
-      if (i != k && numint_sgn(con->p[i][j]))
+      if (i != (size_t)k && numint_sgn(con->p[i][j]))
 	matrix_combine_rows(pk,con,i,(size_t)k,i,j);
     }
   }
